% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apcadrJed_main.R
\name{apcadr}
\alias{apcadr}
\title{apcadr}
\usage{
apcadr(x)
}
\arguments{
\item{x}{Row of a dataset.}
}
\description{
apcadr
}
\note{
This function counts "腺腫" while excluding "家族性大腸腺腫症", "Peutz-Jeghers症候群", "若年性ポリポーシス" and "Serrated polyposis syndrome", according to JED-style database.
You can find sample database in "tests/testthat/sample_CF_data.csv".
}
\examples{
data<-(c("腺腫|腺腫|憩室", "憩室|痔核", "特記所見なし", "家族性大腸腺腫症"))
apcadr(data)

# returns "APC:0.66667, ADR:0.33333"

}
